requirements.txt
absl-py==2.3.1
annotated-types==0.7.0
certifi==2025.10.5
charset-normalizer==3.4.3
click==8.1.8
contourpy==1.3.0
cycler==0.12.1
eval_type_backport==0.2.2
fonttools==4.60.1
ftfy==6.3.1
gitdb==4.0.12
GitPython==3.1.45
grpcio==1.75.1
idna==3.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
kiwisolver==1.4.7
Markdown==3.9
MarkupSafe==3.0.3
matplotlib==3.9.4
ninja==1.13.0
numpy==1.24.3
openai-clip==1.0.1
opencv-python==4.12.0.88
packaging==25.0
pandas==2.3.3
pillow==11.3.0
platformdirs==4.4.0
protobuf==6.32.1
pydantic==2.12.0
pydantic_core==2.41.1
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
regex==2025.9.18
requests==2.32.5
scipy==1.13.1
sentry-sdk==2.40.0
six==1.17.0
smmap==5.0.2
tensorboard==2.20.0
tensorboard-data-server==0.7.2
torch==1.13.0+cu117
torchaudio==0.13.0+cu117
torchvision==0.14.0+cu117
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
wandb==0.22.2
wcwidth==0.2.14
Werkzeug==3.1.3
zipp==3.23.0
